feat: 차량 상세페이지 조회 API 및 mock 데이터 응답 구현
- 차량 상세 페이지용 DTO 구조 적용
- CarService에 상세 mock 데이터 생성 로직 추가
- /api/cars/{carId} 상세 조회 컨트롤러 구현

+++ backend/src/main/java/com/carpick/domain/car/dto/response/common/BranchLocationDto.java
@@ -3,11 +3,13 @@ package com.carpick.domain.car.dto.response.common;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BranchLocationDto {
     private Long branchId;
     private String branchName;
     private String address;
-    private Long latitude;
-    private Long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 }
